remove tabble toggle
